Revert "Champion forge button up"
