+++ server/__tests__/championForge.test.js
@@ -12,9 +12,8 @@ process.env.NODE_ENV = 'test';
 const { Sequelize } = require('sequelize');
 const db = require('../db/models');
 const { Op } = require('sequelize');
-const { Mutation: resolvers, Query: queryResolvers } = require('../schema/resolvers/ClanWars');
+const { Mutation: resolvers } = require('../schema/resolvers/ClanWars');
 const { triggerGatheringTransition } = require('../utils/cwScheduler');
-const { generateId } = require('../utils/cwTaskSampler');
 
 // ── DB setup ──────────────────────────────────────────────────────────────────
 
@@ -337,661 +336,5 @@ describe('Champion Forge — full game flow', () => {
     await ClanWarsTask.destroy({ where: { eventId } });
     await ClanWarsTeam.destroy({ where: { eventId } });
     await ClanWarsEvent.destroy({ where: { eventId } });
-  });
-});
-
-// ── Battle system ─────────────────────────────────────────────────────────────
-
-describe('Champion Forge — battle system', () => {
-  let battleEventId;
-  let redTeamId;
-  let blueTeamId;
-
-  // Item IDs for red team
-  let redWeaponId;
-  let redRareWeaponId;
-  let redConsumableId;
-
-  // Item IDs for blue team
-  let blueWeaponId;
-
-  // Battle under test
-  let battleId;
-
-  // ── Setup ──────────────────────────────────────────────────────────────────
-
-  beforeAll(async () => {
-    const { ClanWarsEvent, ClanWarsTeam, ClanWarsItem } = db;
-
-    // 1. Create event in DRAFT status
-    const event = await resolvers.createClanWarsEvent(
-      null,
-      {
-        input: {
-          eventName: 'Battle System Test Event',
-          difficulty: 'standard',
-          bracketType: 'SINGLE_ELIMINATION',
-          gatheringHours: 1,
-          outfittingHours: 1,
-          flexRolesAllowed: false,
-          teams: [
-            {
-              teamName: 'Red Team',
-              members: [
-                { discordId: adminUser.discordUserId, username: 'Admin', avatar: null, role: 'UNSET' },
-              ],
-            },
-            {
-              teamName: 'Blue Team',
-              members: [
-                { discordId: pvmerUser.discordUserId, username: 'PvMer', avatar: null, role: 'UNSET' },
-              ],
-            },
-          ],
-        },
-      },
-      ctx(adminUser)
-    );
-    battleEventId = event.eventId;
-
-    const teams = await ClanWarsTeam.findAll({ where: { eventId: battleEventId } });
-    redTeamId = teams.find((t) => t.teamName === 'Red Team').teamId;
-    blueTeamId = teams.find((t) => t.teamName === 'Blue Team').teamId;
-
-    // 2. Set captains so startClanWarsBattle can find them
-    await resolvers.setClanWarsCaptain(
-      null,
-      { teamId: redTeamId, discordId: adminUser.discordUserId },
-      ctx(adminUser)
-    );
-    await resolvers.setClanWarsCaptain(
-      null,
-      { teamId: blueTeamId, discordId: pvmerUser.discordUserId },
-      ctx(adminUser)
-    );
-
-    // 3. Create items for red team
-    redWeaponId = generateId('cwi');
-    redRareWeaponId = generateId('cwi');
-    redConsumableId = generateId('cwi');
-
-    await ClanWarsItem.bulkCreate([
-      {
-        itemId: redWeaponId,
-        teamId: redTeamId,
-        eventId: battleEventId,
-        name: 'Test Sword',
-        slot: 'weapon',
-        rarity: 'common',
-        itemSnapshot: {
-          stats: { attack: 12, defense: 0, speed: 8, crit: 3, hp: 0 },
-          spriteKey: null,
-          spriteIcon: null,
-          inventoryIcon: null,
-        },
-        isEquipped: false,
-        isUsed: false,
-        earnedAt: new Date(),
-      },
-      {
-        itemId: redRareWeaponId,
-        teamId: redTeamId,
-        eventId: battleEventId,
-        name: 'Rare Axe',
-        slot: 'weapon',
-        rarity: 'rare',
-        itemSnapshot: {
-          stats: { attack: 32, defense: 0, speed: 15, crit: 14, hp: 0 },
-          special: { id: 'cleave' },
-          spriteKey: null,
-          spriteIcon: null,
-          inventoryIcon: null,
-        },
-        isEquipped: false,
-        isUsed: false,
-        earnedAt: new Date(),
-      },
-      {
-        itemId: redConsumableId,
-        teamId: redTeamId,
-        eventId: battleEventId,
-        name: 'Boar Rib',
-        slot: 'consumable',
-        rarity: 'common',
-        itemSnapshot: {
-          consumableEffect: { type: 'heal', value: 40, description: 'Restores 40 HP.' },
-          spriteKey: null,
-          spriteIcon: null,
-          inventoryIcon: null,
-        },
-        isEquipped: false,
-        isUsed: false,
-        earnedAt: new Date(),
-      },
-    ]);
-
-    // 4. Create items for blue team
-    blueWeaponId = generateId('cwi');
-    await ClanWarsItem.create({
-      itemId: blueWeaponId,
-      teamId: blueTeamId,
-      eventId: battleEventId,
-      name: 'Iron Dagger',
-      slot: 'weapon',
-      rarity: 'common',
-      itemSnapshot: {
-        stats: { attack: 10, defense: 2, speed: 6, crit: 2, hp: 0 },
-        spriteKey: null,
-        spriteIcon: null,
-        inventoryIcon: null,
-      },
-      isEquipped: false,
-      isUsed: false,
-      earnedAt: new Date(),
-    });
-
-    // 5. Advance event to OUTFITTING so lockClanWarsLoadout is allowed
-    await resolvers.adminForceEventStatus(
-      null,
-      { eventId: battleEventId, status: 'OUTFITTING' },
-      ctx(adminUser)
-    );
-
-    // 6. Save and lock loadouts — red team uses rare weapon (has 'cleave') + consumable
-    await resolvers.saveOfficialLoadout(
-      null,
-      {
-        teamId: redTeamId,
-        loadout: {
-          weapon: redRareWeaponId,
-          helm: null,
-          chest: null,
-          legs: null,
-          gloves: null,
-          boots: null,
-          shield: null,
-          ring: null,
-          amulet: null,
-          cape: null,
-          trinket: null,
-          consumables: [redConsumableId],
-          chosenSpecial: 'cleave',
-        },
-      },
-      ctx(adminUser)
-    );
-    await resolvers.lockClanWarsLoadout(null, { teamId: redTeamId }, ctx(adminUser));
-
-    await resolvers.saveOfficialLoadout(
-      null,
-      {
-        teamId: blueTeamId,
-        loadout: {
-          weapon: blueWeaponId,
-          helm: null,
-          chest: null,
-          legs: null,
-          gloves: null,
-          boots: null,
-          shield: null,
-          ring: null,
-          amulet: null,
-          cape: null,
-          trinket: null,
-          consumables: [],
-          chosenSpecial: null,
-        },
-      },
-      ctx(adminUser)
-    );
-    await resolvers.lockClanWarsLoadout(null, { teamId: blueTeamId }, ctx(adminUser));
-
-    // 7. Advance to BATTLE and generate bracket
-    await resolvers.adminForceEventStatus(
-      null,
-      { eventId: battleEventId, status: 'BATTLE' },
-      ctx(adminUser)
-    );
-    await resolvers.generateClanWarsBracket(
-      null,
-      { eventId: battleEventId, bracketType: 'SINGLE_ELIMINATION' },
-      ctx(adminUser)
-    );
-
-    // 8. Start the battle
-    const battle = await resolvers.startClanWarsBattle(
-      null,
-      { eventId: battleEventId, team1Id: redTeamId, team2Id: blueTeamId },
-      ctx(adminUser)
-    );
-    battleId = battle.battleId;
-  }, 30000);
-
-  // ── Core battle actions ────────────────────────────────────────────────────
-
-  test('startClanWarsBattle creates a battle IN_PROGRESS with correct initial state', async () => {
-    const { ClanWarsBattle, ClanWarsBattleEvent: ClanWarsBattleLog } = db;
-
-    const battle = await ClanWarsBattle.findByPk(battleId);
-    expect(battle).not.toBeNull();
-    expect(battle.status).toBe('IN_PROGRESS');
-    expect(battle.team1Id).toBe(redTeamId);
-    expect(battle.team2Id).toBe(blueTeamId);
-
-    const state = battle.battleState;
-    const snap = battle.championSnapshots;
-
-    // HP values match champion maxHp
-    expect(state.hp.team1).toBe(snap.champion1.stats.maxHp);
-    expect(state.hp.team2).toBe(snap.champion2.stats.maxHp);
-
-    // Team with higher speed goes first (red team has speed 15 from rare axe vs blue team speed 6)
-    expect(state.currentTurn).toBe('team1');
-
-    // BATTLE_START log entry exists
-    const log = await ClanWarsBattleLog.findAll({ where: { battleId, action: 'BATTLE_START' } });
-    expect(log).toHaveLength(1);
-    expect(log[0].turnNumber).toBe(0);
-    expect(log[0].actorTeamId).toBeNull();
-  });
-
-  test('ATTACK action reduces defender HP', async () => {
-    const { ClanWarsBattle } = db;
-
-    const battleBefore = await ClanWarsBattle.findByPk(battleId);
-    const stateBefore = battleBefore.battleState;
-    const actorSide = stateBefore.currentTurn;
-    const defSide = actorSide === 'team1' ? 'team2' : 'team1';
-    const actorTeamId = actorSide === 'team1' ? redTeamId : blueTeamId;
-    const hpBefore = stateBefore.hp[defSide];
-
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: actorTeamId, action: 'ATTACK' },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    const stateAfter = battleAfter.battleState;
-
-    expect(stateAfter.hp[defSide]).toBeLessThan(hpBefore);
-    expect(stateAfter.hp[defSide]).toBeGreaterThanOrEqual(0);
-  });
-
-  test('turn advances to the other team after each action', async () => {
-    const { ClanWarsBattle } = db;
-
-    const battleBefore = await ClanWarsBattle.findByPk(battleId);
-    const turnBefore = battleBefore.battleState.currentTurn;
-    const actorTeamId = turnBefore === 'team1' ? redTeamId : blueTeamId;
-
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: actorTeamId, action: 'ATTACK' },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    const turnAfter = battleAfter.battleState.currentTurn;
-
-    const expected = turnBefore === 'team1' ? 'team2' : 'team1';
-    expect(turnAfter).toBe(expected);
-  });
-
-  test('non-turn-holder cannot submit an action', async () => {
-    const { ClanWarsBattle } = db;
-
-    const battle = await ClanWarsBattle.findByPk(battleId);
-    const currentTurn = battle.battleState.currentTurn;
-    // The team whose turn it is NOT
-    const wrongTeamId = currentTurn === 'team1' ? blueTeamId : redTeamId;
-
-    await expect(
-      resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: wrongTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      )
-    ).rejects.toThrow('not your turn');
-  });
-
-  test('DEFEND action sets defendActive flag for acting team', async () => {
-    const { ClanWarsBattle } = db;
-
-    const battleBefore = await ClanWarsBattle.findByPk(battleId);
-    const actorSide = battleBefore.battleState.currentTurn;
-    const actorTeamId = actorSide === 'team1' ? redTeamId : blueTeamId;
-
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: actorTeamId, action: 'DEFEND' },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    expect(battleAfter.battleState.defendActive[actorSide]).toBe(true);
-  });
-
-  test('attacking a defending champion deals reduced damage', async () => {
-    const { ClanWarsBattle } = db;
-
-    // The defender side is the team that just DEFENDed (actorSide from previous test).
-    // After DEFEND, it's now the other team's turn.
-    const battleNow = await ClanWarsBattle.findByPk(battleId);
-    const state = battleNow.battleState;
-    const actorSide = state.currentTurn;
-    const defSide = actorSide === 'team1' ? 'team2' : 'team1';
-    const actorTeamId = actorSide === 'team1' ? redTeamId : blueTeamId;
-
-    // Only meaningful if the defender is actually defending
-    const defenderIsDefending = state.defendActive[defSide];
-    const hpBefore = state.hp[defSide];
-
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: actorTeamId, action: 'ATTACK' },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    const hpAfter = battleAfter.battleState.hp[defSide];
-    const damageDealt = hpBefore - hpAfter;
-
-    if (defenderIsDefending) {
-      // Damage with defend active is at 40% of normal; with attack=32, defense=0:
-      // base = 32, with defend: ~32 * 0.85-1.15 * 0.4 ≈ 11-15 range. Without defend: ~27-37.
-      // At minimum the damage should be at least 1 and at most ~20 (40% cap).
-      expect(damageDealt).toBeGreaterThanOrEqual(1);
-      expect(damageDealt).toBeLessThanOrEqual(25);
-    } else {
-      // Defender wasn't defending; just verify damage > 0
-      expect(damageDealt).toBeGreaterThanOrEqual(1);
-    }
-  });
-
-  test('SPECIAL action fires and marks specialUsed for that team', async () => {
-    const { ClanWarsBattle } = db;
-
-    // Find whose turn it is; red team (team1) has 'cleave' special
-    // We need it to be red team's turn. Advance turns with ATTACKs until it's team1's turn.
-    let battle = await ClanWarsBattle.findByPk(battleId);
-    while (battle.battleState.currentTurn !== 'team1' && battle.status === 'IN_PROGRESS') {
-      await resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: blueTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      );
-      battle = await ClanWarsBattle.findByPk(battleId);
-    }
-
-    // Guard: if battle ended before we could test special, skip gracefully
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Red team uses their SPECIAL (cleave)
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: redTeamId, action: 'SPECIAL' },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    expect(battleAfter.battleState.specialUsed.team1).toBe(true);
-  });
-
-  test('using a special twice throws an error containing "already used"', async () => {
-    const { ClanWarsBattle } = db;
-
-    let battle = await ClanWarsBattle.findByPk(battleId);
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Advance to team1's turn again
-    while (battle.battleState.currentTurn !== 'team1' && battle.status === 'IN_PROGRESS') {
-      await resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: blueTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      );
-      battle = await ClanWarsBattle.findByPk(battleId);
-    }
-
-    if (battle.status !== 'IN_PROGRESS') return;
-    if (!battle.battleState.specialUsed.team1) return; // special not yet used — precondition not met
-
-    await expect(
-      resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: redTeamId, action: 'SPECIAL' },
-        ctx(adminUser)
-      )
-    ).rejects.toThrow(/already used/i);
-  });
-
-  test('USE_ITEM heal restores HP and removes item from consumablesRemaining', async () => {
-    const { ClanWarsBattle } = db;
-
-    let battle = await ClanWarsBattle.findByPk(battleId);
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Manually inject the consumable into consumablesRemaining for team1 so USE_ITEM works.
-    // (initBattleState uses snap.consumables which is not set by startClanWarsBattle — this
-    //  patches the live battleState to simulate the item being available.)
-    const patchedState = {
-      ...battle.battleState,
-      consumablesRemaining: {
-        ...battle.battleState.consumablesRemaining,
-        team1: [redConsumableId],
-      },
-    };
-    await battle.update({ battleState: patchedState });
-
-    // Deal some damage first (advance to team2's turn, attack team1, then back to team1's turn)
-    battle = await ClanWarsBattle.findByPk(battleId);
-    while (battle.battleState.currentTurn !== 'team2' && battle.status === 'IN_PROGRESS') {
-      await resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: redTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      );
-      battle = await ClanWarsBattle.findByPk(battleId);
-    }
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Blue team attacks red team to deal some damage
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: blueTeamId, action: 'ATTACK' },
-      ctx(adminUser)
-    );
-    battle = await ClanWarsBattle.findByPk(battleId);
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Advance to team1's turn
-    while (battle.battleState.currentTurn !== 'team1' && battle.status === 'IN_PROGRESS') {
-      await resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: blueTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      );
-      battle = await ClanWarsBattle.findByPk(battleId);
-    }
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Ensure consumable is still in the remaining list (might have been cleared by turn advances)
-    const stateCheck = battle.battleState;
-    if (!stateCheck.consumablesRemaining.team1.includes(redConsumableId)) {
-      await battle.update({
-        battleState: {
-          ...stateCheck,
-          consumablesRemaining: { ...stateCheck.consumablesRemaining, team1: [redConsumableId] },
-        },
-      });
-    }
-
-    battle = await ClanWarsBattle.findByPk(battleId);
-    const hpBeforeHeal = battle.battleState.hp.team1;
-    const maxHp = battle.championSnapshots.champion1.stats.maxHp;
-
-    await resolvers.submitBattleAction(
-      null,
-      { battleId, teamId: redTeamId, action: 'USE_ITEM', itemId: redConsumableId },
-      ctx(adminUser)
-    );
-
-    const battleAfter = await ClanWarsBattle.findByPk(battleId);
-    const hpAfterHeal = battleAfter.battleState.hp.team1;
-
-    // HP should have gone up (capped at maxHp) unless already at max
-    if (hpBeforeHeal < maxHp) {
-      expect(hpAfterHeal).toBeGreaterThan(hpBeforeHeal);
-    }
-    expect(hpAfterHeal).toBeLessThanOrEqual(maxHp);
-
-    // Consumable should be removed from consumablesRemaining
-    expect(battleAfter.battleState.consumablesRemaining.team1).not.toContain(redConsumableId);
-  });
-
-  test('USE_ITEM with an invalid itemId throws an error', async () => {
-    const { ClanWarsBattle } = db;
-
-    let battle = await ClanWarsBattle.findByPk(battleId);
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    // Advance to team1's turn
-    while (battle.battleState.currentTurn !== 'team1' && battle.status === 'IN_PROGRESS') {
-      await resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: blueTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      );
-      battle = await ClanWarsBattle.findByPk(battleId);
-    }
-    if (battle.status !== 'IN_PROGRESS') return;
-
-    const fakeItemId = generateId('cwi');
-    await expect(
-      resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: redTeamId, action: 'USE_ITEM', itemId: fakeItemId },
-        ctx(adminUser)
-      )
-    ).rejects.toThrow(/not available|not found/i);
-  });
-
-  test('devAutoBattle completes the battle with a winner and BATTLE_END log entry', async () => {
-    const { ClanWarsBattle, ClanWarsBattleEvent: ClanWarsBattleLog } = db;
-
-    let battle = await ClanWarsBattle.findByPk(battleId);
-    if (battle.status === 'COMPLETED') {
-      // Battle may have already ended during earlier action tests — verify directly
-      expect(battle.winnerId).not.toBeNull();
-      const endLog = await ClanWarsBattleLog.findAll({
-        where: { battleId, action: 'BATTLE_END' },
-      });
-      expect(endLog.length).toBeGreaterThanOrEqual(1);
-      return;
-    }
-
-    const completedBattle = await resolvers.devAutoBattle(
-      null,
-      { battleId },
-      ctx(adminUser)
-    );
-
-    expect(completedBattle.status).toBe('COMPLETED');
-    expect(completedBattle.winnerId).not.toBeNull();
-    expect([redTeamId, blueTeamId]).toContain(completedBattle.winnerId);
-
-    // BATTLE_END log entry exists
-    const endLog = await ClanWarsBattleLog.findAll({
-      where: { battleId, action: 'BATTLE_END' },
-    });
-    expect(endLog.length).toBeGreaterThanOrEqual(1);
-  });
-
-  test('getClanWarsBattleLog returns log entries including BATTLE_START', async () => {
-    const log = await queryResolvers.getClanWarsBattleLog(null, { battleId });
-    expect(Array.isArray(log)).toBe(true);
-    expect(log.length).toBeGreaterThan(0);
-
-    const startEntry = log.find((e) => e.action === 'BATTLE_START');
-    expect(startEntry).toBeDefined();
-    expect(startEntry.battleId).toBe(battleId);
-  });
-
-  test('bracket advances after battle completes — winnerId set in bracket match', async () => {
-    const { ClanWarsEvent, ClanWarsBattle } = db;
-
-    const battle = await ClanWarsBattle.findByPk(battleId);
-    expect(battle.status).toBe('COMPLETED');
-    expect(battle.winnerId).not.toBeNull();
-
-    const event = await ClanWarsEvent.findByPk(battleEventId);
-    const bracket = event.bracket;
-    expect(bracket).not.toBeNull();
-
-    // Find the match with our battleId in any bracket section
-    const allMatches = [
-      ...(bracket.rounds ?? []).flatMap((r) => r.matches),
-      ...(bracket.losersBracket ?? []).flatMap((r) => r.matches),
-    ];
-    if (bracket.grandFinal) allMatches.push(bracket.grandFinal);
-
-    const match = allMatches.find((m) => m.battleId === battleId);
-    expect(match).toBeDefined();
-    expect(match.winnerId).toBe(battle.winnerId);
-  });
-
-  test('devSimulateNextMatch starts the next bracket match (or event is completed if only one match)', async () => {
-    const { ClanWarsEvent } = db;
-
-    const event = await ClanWarsEvent.findByPk(battleEventId);
-    // With 2 teams in SINGLE_ELIMINATION there is only one match (already played).
-    // devSimulateNextMatch should throw "No unstarted matches" OR the event has auto-completed.
-    if (event.status === 'COMPLETED') {
-      // Correct: all matches done, event auto-completed
-      expect(event.status).toBe('COMPLETED');
-      return;
-    }
-
-    // If there are unstarted matches (would happen with > 2 teams), simulate one
-    await expect(
-      resolvers.devSimulateNextMatch(null, { eventId: battleEventId }, ctx(adminUser))
-    ).rejects.toThrow(/no unstarted/i);
-  });
-
-  test('cannot submit action to a completed battle', async () => {
-    const { ClanWarsBattle } = db;
-    const battle = await ClanWarsBattle.findByPk(battleId);
-    expect(battle.status).toBe('COMPLETED');
-
-    await expect(
-      resolvers.submitBattleAction(
-        null,
-        { battleId, teamId: redTeamId, action: 'ATTACK' },
-        ctx(adminUser)
-      )
-    ).rejects.toThrow(/not in progress/i);
-  });
-
-  // ── Cleanup ────────────────────────────────────────────────────────────────
-
-  afterAll(async () => {
-    if (!battleEventId) return;
-    const {
-      ClanWarsEvent,
-      ClanWarsTeam,
-      ClanWarsTask,
-      ClanWarsSubmission,
-      ClanWarsItem,
-      ClanWarsBattle,
-      ClanWarsBattleEvent,
-    } = db;
-    await ClanWarsBattleEvent.destroy({ where: { battleId } });
-    await ClanWarsBattle.destroy({ where: { eventId: battleEventId } });
-    await ClanWarsSubmission.destroy({ where: { eventId: battleEventId } });
-    await ClanWarsItem.destroy({ where: { eventId: battleEventId } });
-    await ClanWarsTask.destroy({ where: { eventId: battleEventId } });
-    await ClanWarsTeam.destroy({ where: { eventId: battleEventId } });
-    await ClanWarsEvent.destroy({ where: { eventId: battleEventId } });
   });
 });
