+++ server/db/config/config.json
@@ -21,12 +21,6 @@
         "require": true,
         "rejectUnauthorized": false
       }
-    },
-    "pool": {
-      "max": 8,
-      "min": 0,
-      "acquire": 30000,
-      "idle": 10000
     }
   },
   "production": {
@@ -38,12 +32,6 @@
         "require": true,
         "rejectUnauthorized": false
       }
-    },
-    "pool": {
-      "max": 8,
-      "min": 0,
-      "acquire": 30000,
-      "idle": 10000
     }
   }
 }
