+++ server/db/db.js
@@ -8,10 +8,9 @@ const sharedOptions = {
   dialect: 'postgres',
   benchmark: true,
   logging: isDev ? (sql, timing) => logger.debug({ timing, sql }, 'SQL') : false,
-  // This instance is only used for sync/authenticate at startup — keep it tiny
   pool: {
-    max: 3,
-    min: 0,
+    max: 20,
+    min: 2,
     acquire: 30000,
     idle: 10000,
   },
