+++ server/package.json
@@ -33,13 +33,11 @@
     "express-session": "^1.18.1",
     "fuse.js": "^7.0.0",
     "graphql": "^15.9.0",
-    "graphql-redis-subscriptions": "^2.7.0",
     "graphql-scalars": "^1.24.2",
     "graphql-subscriptions": "^3.0.0",
     "graphql-tag": "^2.12.6",
     "graphql-ws": "^6.0.6",
     "helmet": "^8.1.0",
-    "ioredis": "^5.11.1",
     "jsonwebtoken": "^9.0.2",
     "node-cache": "^5.1.2",
     "node-cron": "^4.2.1",
