+++ server/utils/cwObjectiveCollections.js
@@ -23,22 +23,20 @@ const PVM_INITIATE = [
     descriptionTemplate: 'Obtain {quantity} drops from Amoxliatl.',
     acceptableItems: ['Moxi', 'Glacial Temotli'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
     id: 'pvm_bryophyta',
-    label: "Moss Giant's Bane",
+    label: 'Moss Giant\'s Bane',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'initiate',
     boss: 'Bryophyta',
     descriptionTemplate: 'Obtain {quantity} drops from Bryophyta.',
-    acceptableItems: ["Bryophyta's Essence", "Bryophyta's Staff", 'Mossy Key'],
+    acceptableItems: ['Bryophyta\'s Essence', 'Bryophyta\'s Staff', 'Mossy Key'],
     quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -53,9 +51,7 @@ const PVM_INITIATE = [
     descriptionTemplate: 'Obtain {quantity} drops from Kraken.',
     acceptableItems: ['Kraken Tentacle', 'Trident of the Seas', 'Jar of Dirt', 'Pet Kraken'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -68,7 +64,7 @@ const PVM_INITIATE = [
     descriptionTemplate: 'Obtain {quantity} drops from Obor.',
     acceptableItems: ['Hill Giant Club', 'Giant Key'],
     quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -81,16 +77,9 @@ const PVM_INITIATE = [
     difficulty: 'initiate',
     boss: 'Royal Titans',
     descriptionTemplate: 'Obtain {quantity} drops from Royal Titans.',
-    acceptableItems: [
-      'Ice Element Staff Crown',
-      'Fire Element Staff Crown',
-      'Giantsoul Amulet',
-      'Bran',
-    ],
+    acceptableItems: ['Ice Element Staff Crown', 'Fire Element Staff Crown', 'Giantsoul Amulet', 'Bran'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -101,11 +90,9 @@ const PVM_INITIATE = [
     difficulty: 'initiate',
     boss: 'Sarachnis',
     descriptionTemplate: 'Obtain {quantity} drops from Sarachnis.',
-    acceptableItems: ['Sraracha', 'Sarachnis Cudgel', 'Jar of Eyes', 'Pristine Spider Silk'],
+    acceptableItems: ['Sraracha', 'Sarachnis Cudgel', 'Jar of Eyes'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 2, max: 3 },
     },
   },
 ];
@@ -119,17 +106,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Abyssal Sire',
     descriptionTemplate: 'Obtain {quantity} drops from Abyssal Sire.',
-    acceptableItems: [
-      'Abyssal Whip',
-      'Abyssal Dagger',
-      'Abyssal Bludgeon Piece',
-      'Abyssal Orphan',
-      'Jar of Miasma',
-    ],
+    acceptableItems: ['Abyssal Whip', 'Abyssal Dagger', 'Abyssal Bludgeon Piece', 'Abyssal Orphan', 'Jar of Miasma'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -142,9 +121,7 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Araxxor.',
     acceptableItems: ['Nid', 'Araxyte Fang', 'Noxious Point', 'Noxious Blade', 'Noxious Pommel'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -155,18 +132,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Artio',
     descriptionTemplate: 'Obtain {quantity} drops from Artio.',
-    acceptableItems: [
-      'Callisto Cub',
-      'Voidwaker Hilt',
-      'Tyrannical Ring',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-      'Claws of Callisto',
-    ],
+    acceptableItems: ['Callisto Cub', 'Voidwaker Hilt', 'Tyrannical Ring', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -177,58 +145,22 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Barrows',
     descriptionTemplate: 'Obtain {quantity} drops from Barrows.',
-    acceptableItems: [
-      "Ahrim's Hood",
-      "Ahrim's Robetop",
-      "Ahrim's Robeskirt",
-      "Ahrim's Staff",
-      "Dharok's Helm",
-      "Dharok's Platebody",
-      "Dharok's Platelegs",
-      "Dharok's Greataxe",
-      "Guthan's Helm",
-      "Guthan's Platebody",
-      "Guthan's Chainskirt",
-      "Guthan's Warspear",
-      "Karil's Coif",
-      "Karil's Leathertop",
-      "Karil's Leatherskirt",
-      "Karil's Crossbow",
-      "Torag's Helm",
-      "Torag's Platebody",
-      "Torag's Platelegs",
-      "Torag's Hammers",
-      "Verac's Helm",
-      "Verac's Brassard",
-      "Verac's Plateskirt",
-      "Verac's Flail",
-    ],
+    acceptableItems: ['Ahrim\'s Hood', 'Ahrim\'s Robetop', 'Ahrim\'s Robeskirt', 'Ahrim\'s Staff', 'Dharok\'s Helm', 'Dharok\'s Platebody', 'Dharok\'s Platelegs', 'Dharok\'s Greataxe', 'Guthan\'s Helm', 'Guthan\'s Platebody', 'Guthan\'s Chainskirt', 'Guthan\'s Warspear', 'Karil\'s Coif', 'Karil\'s Leathertop', 'Karil\'s Leatherskirt', 'Karil\'s Crossbow', 'Torag\'s Helm', 'Torag\'s Platebody', 'Torag\'s Platelegs', 'Torag\'s Hammers', 'Verac\'s Helm', 'Verac\'s Brassard', 'Verac\'s Plateskirt', 'Verac\'s Flail'],
     quantities: {
-      casual: { min: 3, max: 5 },
       standard: { min: 5, max: 7 },
-      hardcore: { min: 7, max: 10 },
     },
   },
   {
     id: 'pvm_callisto',
-    label: "Honey, I'm Home",
+    label: 'Honey, I\'m Home',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
     boss: 'Callisto',
     descriptionTemplate: 'Obtain {quantity} drops from Callisto.',
-    acceptableItems: [
-      'Callisto Cub',
-      'Voidwaker Hilt',
-      'Tyrannical Ring',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-      'Claws of Callisto',
-    ],
+    acceptableItems: ['Callisto Cub', 'Voidwaker Hilt', 'Tyrannical Ring', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -237,20 +169,11 @@ const PVM_ADEPT = [
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
-    boss: "Calvar'ion",
-    descriptionTemplate: "Obtain {quantity} drops from Calvar'ion.",
-    acceptableItems: [
-      "Vet'ion Jr.",
-      'Voidwaker Blade',
-      'Ring of the Gods',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-      "Skull of Vet'ion",
-    ],
+    boss: 'Calvar\'ion',
+    descriptionTemplate: 'Obtain {quantity} drops from Calvar\'ion.',
+    acceptableItems: ['Vet\'ion Jr.', 'Voidwaker Blade', 'Ring of the Gods', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -261,25 +184,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Chambers of Xeric',
     descriptionTemplate: 'Obtain {quantity} drops from Chambers of Xeric.',
-    acceptableItems: [
-      'Olmlet',
-      'Twisted Bow',
-      'Elder Maul',
-      'Kodai Insignia',
-      'Dragon Claws',
-      'Ancestral Hat',
-      'Ancestral Robe Top',
-      'Ancestral Robe Bottom',
-      'Dexterous Prayer Scroll',
-      'Arcane Prayer Scroll',
-      "Dinh's Bulwark",
-      'Dragon Hunter Crossbow',
-      'Twisted Buckler',
-    ],
+    acceptableItems: ['Olmlet', 'Twisted Bow', 'Elder Maul', 'Kodai Insignia', 'Dragon Claws', 'Ancestral Hat', 'Ancestral Robe Top', 'Ancestral Robe Bottom', 'Dexterous Prayer Scroll', 'Arcane Prayer Scroll', 'Dinh\'s Bulwark', 'Dragon Hunter Crossbow', 'Twisted Buckler'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
@@ -290,27 +197,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Challenge Mode Chambers of Xeric',
     descriptionTemplate: 'Obtain {quantity} drops from Challenge Mode Chambers of Xeric.',
-    acceptableItems: [
-      'Olmlet',
-      'Twisted Bow',
-      'Elder Maul',
-      'Kodai Insignia',
-      'Dragon Claws',
-      'Ancestral Hat',
-      'Ancestral Robe Top',
-      'Ancestral Robe Bottom',
-      'Dexterous Prayer Scroll',
-      'Arcane Prayer Scroll',
-      "Dinh's Bulwark",
-      'Dragon Hunter Crossbow',
-      'Twisted Buckler',
-      'Twisted Ancestral Colour Kit',
-      'Metamorphic Dust',
-    ],
+    acceptableItems: ['Olmlet', 'Twisted Bow', 'Elder Maul', 'Kodai Insignia', 'Dragon Claws', 'Ancestral Hat', 'Ancestral Robe Top', 'Ancestral Robe Bottom', 'Dexterous Prayer Scroll', 'Arcane Prayer Scroll', 'Dinh\'s Bulwark', 'Dragon Hunter Crossbow', 'Twisted Buckler', 'Twisted Ancestral Colour Kit', 'Metamorphic Dust'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
@@ -323,9 +212,7 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Chaos Elemental.',
     acceptableItems: ['Pet Chaos Elemental', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -338,9 +225,7 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Chaos Fanatic.',
     acceptableItems: ['Pet Chaos Fanatic', 'Odium Shard', 'Malediction Shard'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -351,17 +236,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Commander Zilyana',
     descriptionTemplate: 'Obtain {quantity} drops from Commander Zilyana.',
-    acceptableItems: [
-      'Saradomin Sword',
-      'Armadyl Crossbow',
-      "Saradomin's Light",
-      'Saradomin Hilt',
-      'Zilyana Jr.',
-    ],
+    acceptableItems: ['Saradomin Sword', 'Armadyl Crossbow', 'Saradomin\'s Light', 'Saradomin Hilt', 'Zilyana Jr.'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -374,9 +251,7 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Crazy Archaeologist.',
     acceptableItems: ['Pet Chaos Elemental', 'Odium Shard', 'Malediction Shard'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -387,22 +262,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Dagannoth Kings',
     descriptionTemplate: 'Obtain {quantity} drops from Dagannoth Kings.',
-    acceptableItems: [
-      'Pet Dagannoth Prime',
-      'Pet Dagannoth Supreme',
-      'Pet Dagannoth Rex',
-      'Berserker Ring',
-      'Archers Ring',
-      'Seers Ring',
-      'Warrior Ring',
-      'Dragon Axe',
-      'Seercull',
-      'Mud Battlestaff',
-    ],
+    acceptableItems: ['Pet Dagannoth Prime', 'Pet Dagannoth Supreme', 'Pet Dagannoth Rex', 'Berserker Ring', 'Archers Ring', 'Seers Ring', 'Warrior Ring', 'Dragon Axe', 'Seercull', 'Mud Battlestaff'],
     quantities: {
-      casual: { min: 2, max: 3 },
       standard: { min: 3, max: 4 },
-      hardcore: { min: 4, max: 5 },
     },
   },
   {
@@ -413,17 +275,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'General Graardor',
     descriptionTemplate: 'Obtain {quantity} drops from General Graardor.',
-    acceptableItems: [
-      'Bandos Chestplate',
-      'Bandos Tassets',
-      'Bandos Boots',
-      'Bandos Hilt',
-      'General Graardor Jr.',
-    ],
+    acceptableItems: ['Bandos Chestplate', 'Bandos Tassets', 'Bandos Boots', 'Bandos Hilt', 'General Graardor Jr.'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -434,19 +288,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Grotesque Guardians',
     descriptionTemplate: 'Obtain {quantity} drops from Grotesque Guardians.',
-    acceptableItems: [
-      'Noon',
-      'Granite Maul',
-      'Granite Gloves',
-      'Granite Ring',
-      'Granite Hammer',
-      'Black Tourmaline Core',
-      'Jar of Stone',
-    ],
+    acceptableItems: ['Noon', 'Granite Maul', 'Granite Gloves', 'Granite Ring', 'Granite Hammer', 'Black Tourmaline Core', 'Jar of Stone'],
     quantities: {
-      casual: { min: 2, max: 2 },
-      standard: { min: 3, max: 3 },
-      hardcore: { min: 3, max: 4 },
+      standard: { min: 3, max: 4 },
     },
   },
   {
@@ -459,45 +303,20 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Hueycoatl.',
     acceptableItems: ['Huberte', 'Hueycoatl Hide', 'Tome of Earth', 'Dragon Hunter Wand'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
-    },
-  },
-  {
-    id: 'pvm_giantMole',
-    label: 'Underground Menace',
-    type: 'item_collection',
-    role: 'PVMER',
-    difficulty: 'adept',
-    boss: 'Giant Mole',
-    descriptionTemplate: 'Obtain {quantity} drops from the Giant Mole.',
-    acceptableItems: ['Baby Mole', 'Immaculate Mole Skin'],
-    quantities: {
-      casual: { min: 4, max: 6 },
-      standard: { min: 6, max: 10 },
-      hardcore: { min: 10, max: 14 },
     },
   },
   {
     id: 'pvm_kalphiteQueen',
-    label: "Queen's Gambit",
+    label: 'Queen\'s Gambit',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
     boss: 'Kalphite Queen',
     descriptionTemplate: 'Obtain {quantity} drops from Kalphite Queen.',
-    acceptableItems: [
-      'Kalphite Princess',
-      'Dragon Chainbody',
-      'Dragon 2h Sword',
-      'Jar of Sand',
-      'Kq Head',
-    ],
+    acceptableItems: ['Kalphite Princess', 'Dragon Chainbody', 'Dragon 2h Sword', 'Jar of Sand', 'Kq Head'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -506,40 +325,24 @@ const PVM_ADEPT = [
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
-    boss: "Kree'Arra",
-    descriptionTemplate: "Obtain {quantity} drops from Kree'Arra.",
-    acceptableItems: [
-      'Armadyl Chestplate',
-      'Armadyl Chainskirt',
-      'Armadyl Helmet',
-      'Armadyl Hilt',
-      "Kree'arra Jr.",
-    ],
+    boss: 'Kree\'Arra',
+    descriptionTemplate: 'Obtain {quantity} drops from Kree\'Arra.',
+    acceptableItems: ['Armadyl Chestplate', 'Armadyl Chainskirt', 'Armadyl Helmet', 'Armadyl Hilt', 'Kree\'arra Jr.'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
     id: 'pvm_krilTsutsaroth',
-    label: "Zamorak's Champion",
+    label: 'Zamorak\'s Champion',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
-    boss: "K'ril Tsutsaroth",
-    descriptionTemplate: "Obtain {quantity} drops from K'ril Tsutsaroth.",
-    acceptableItems: [
-      'Staff of the Dead',
-      'Zamorakian Spear',
-      'Steam Battlestaff',
-      'Zamorak Hilt',
-      "K'ril Tsutsaroth Jr.",
-    ],
+    boss: 'K\'ril Tsutsaroth',
+    descriptionTemplate: 'Obtain {quantity} drops from K\'ril Tsutsaroth.',
+    acceptableItems: ['Staff of the Dead', 'Zamorakian Spear', 'Steam Battlestaff', 'Zamorak Hilt', 'K\'ril Tsutsaroth Jr.'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -550,24 +353,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Moons of Peril',
     descriptionTemplate: 'Obtain {quantity} drops from Moons of Peril.',
-    acceptableItems: [
-      'Blood Moon Helm',
-      'Blood Moon Chestplate',
-      'Blood Moon Tassets',
-      'Blue Moon Helm',
-      'Blue Moon Chestplate',
-      'Blue Moon Tassets',
-      'Eclipse Moon Helm',
-      'Eclipse Moon Chestplate',
-      'Eclipse Moon Tassets',
-      'Dual Macuahuitl',
-      'Blue Moon Spear',
-      'Eclipse Atlatl',
-    ],
+    acceptableItems: ['Blood Moon Helm', 'Blood Moon Chestplate', 'Blood Moon Tassets', 'Blue Moon Helm', 'Blue Moon Chestplate', 'Blue Moon Tassets', 'Eclipse Moon Helm', 'Eclipse Moon Chestplate', 'Eclipse Moon Tassets', 'Dual Macuahuitl'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -580,9 +368,7 @@ const PVM_ADEPT = [
     descriptionTemplate: 'Obtain {quantity} drops from Phantom Muspah.',
     acceptableItems: ['Muphin', 'Ancient Icon', 'Venator Shard'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -593,11 +379,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Scorpia',
     descriptionTemplate: 'Obtain {quantity} drops from Scorpia.',
-    acceptableItems: ["Scorpia's Offspring", 'Odium Shard', 'Malediction Shard'],
+    acceptableItems: ['Scorpia\'s Offspring', 'Odium Shard', 'Malediction Shard'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -608,11 +392,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Shellbane Gryphon',
     descriptionTemplate: 'Obtain {quantity} drops from Shellbane Gryphon.',
-    acceptableItems: ['Gull', "Belle's Folly", 'Jar of Feathers'],
+    acceptableItems: ['Gull', 'Belle\'s Folly', 'Jar of Feathers'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -623,18 +405,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Spindel',
     descriptionTemplate: 'Obtain {quantity} drops from Spindel.',
-    acceptableItems: [
-      'Venenatis Spiderling',
-      'Voidwaker Gem',
-      'Treasonous Ring',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-      'Fangs of Venenatis',
-    ],
+    acceptableItems: ['Venenatis Spiderling', 'Voidwaker Gem', 'Treasonous Ring', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -645,20 +418,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Theatre of Blood',
     descriptionTemplate: 'Obtain {quantity} drops from Theatre of Blood.',
-    acceptableItems: [
-      "Lil' Zik",
-      'Scythe of Vitur',
-      'Ghrazi Rapier',
-      'Sanguinesti Staff',
-      'Avernic Defender Hilt',
-      'Justiciar Faceguard',
-      'Justiciar Chestguard',
-      'Justiciar Legguards',
-    ],
+    acceptableItems: ['Lil\' Zik', 'Scythe of Vitur', 'Ghrazi Rapier', 'Sanguinesti Staff', 'Avernic Defender Hilt', 'Justiciar Faceguard', 'Justiciar Chestguard', 'Justiciar Legguards'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
@@ -669,43 +431,22 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Hard Mode Theatre of Blood',
     descriptionTemplate: 'Obtain {quantity} drops from Hard Mode Theatre of Blood.',
-    acceptableItems: [
-      "Lil' Zik",
-      'Scythe of Vitur',
-      'Ghrazi Rapier',
-      'Sanguinesti Staff',
-      'Avernic Defender Hilt',
-      'Justiciar Faceguard',
-      'Justiciar Chestguard',
-      'Justiciar Legguards',
-      'Holy Ornament Kit',
-      'Sanguine Ornament Kit',
-      'Sanguine Dust',
-    ],
+    acceptableItems: ['Lil\' Zik', 'Scythe of Vitur', 'Ghrazi Rapier', 'Sanguinesti Staff', 'Avernic Defender Hilt', 'Justiciar Faceguard', 'Justiciar Chestguard', 'Justiciar Legguards', 'Holy Ornament Kit', 'Sanguine Ornament Kit', 'Sanguine Dust'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
     id: 'pvm_thermonuclearSmokeDevil',
-    label: "Where There's Smoke",
+    label: 'Where There\'s Smoke',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
     boss: 'Thermonuclear Smoke Devil',
     descriptionTemplate: 'Obtain {quantity} drops from Thermonuclear Smoke Devil.',
-    acceptableItems: [
-      'Pet Smoke Devil',
-      'Occult Necklace',
-      'Smoke Battlestaff',
-      'Dragon Chainbody',
-    ],
+    acceptableItems: ['Pet Smoke Devil', 'Occult Necklace', 'Smoke Battlestaff', 'Dragon Chainbody'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -716,20 +457,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Tombs of Amascut',
     descriptionTemplate: 'Obtain {quantity} drops from Tombs of Amascut.',
-    acceptableItems: [
-      "Tumeken's Guardian",
-      "Tumeken's Shadow",
-      "Osmumten's Fang",
-      'Masori Mask',
-      'Masori Body',
-      'Masori Chaps',
-      'Lightbearer',
-      "Elidinis' Ward",
-    ],
+    acceptableItems: ['Tumeken\'s Guardian', 'Tumeken\'s Shadow', 'Osmumten\'s Fang', 'Masori Mask', 'Masori Body', 'Masori Chaps', 'Lightbearer', 'Elidinis\' Ward'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
@@ -740,38 +470,22 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Venenatis',
     descriptionTemplate: 'Obtain {quantity} drops from Venenatis.',
-    acceptableItems: [
-      'Venenatis Spiderling',
-      'Voidwaker Gem',
-      'Treasonous Ring',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-    ],
+    acceptableItems: ['Venenatis Spiderling', 'Voidwaker Gem', 'Treasonous Ring', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
     id: 'pvm_vetion',
-    label: "Death's Hound",
+    label: 'Death\'s Hound',
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'adept',
-    boss: "Vet'ion",
-    descriptionTemplate: "Obtain {quantity} drops from Vet'ion.",
-    acceptableItems: [
-      "Vet'ion Jr.",
-      'Voidwaker Blade',
-      'Ring of the Gods',
-      'Dragon 2h Sword',
-      'Dragon Pickaxe',
-    ],
+    boss: 'Vet\'ion',
+    descriptionTemplate: 'Obtain {quantity} drops from Vet\'ion.',
+    acceptableItems: ['Vet\'ion Jr.', 'Voidwaker Blade', 'Ring of the Gods', 'Dragon 2h Sword', 'Dragon Pickaxe'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -782,18 +496,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Vorkath',
     descriptionTemplate: 'Obtain {quantity} drops from Vorkath.',
-    acceptableItems: [
-      'Dragonbone Necklace',
-      'Jar of Decay',
-      "Vorkath's Head",
-      'Vorki',
-      'Skeletal Visage',
-      'Draconic Visage',
-    ],
+    acceptableItems: ['Dragonbone Necklace', 'Jar of Decay', 'Vorkath\'s Head', 'Vorki', 'Skeletal Visage', 'Draconic Visage'],
     quantities: {
-      casual: { min: 2, max: 3 },
       standard: { min: 3, max: 4 },
-      hardcore: { min: 4, max: 5 },
     },
   },
   {
@@ -804,20 +509,9 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     boss: 'Zulrah',
     descriptionTemplate: 'Obtain {quantity} drops from Zulrah.',
-    acceptableItems: [
-      'Tanzanite Fang',
-      'Magic Fang',
-      'Serpentine Visage',
-      'Uncut Onyx',
-      'Snakeling',
-      'Jar of Swamp',
-      'Tanzanite Mutagen',
-      'Magma Mutagen',
-    ],
+    acceptableItems: ['Tanzanite Fang', 'Magic Fang', 'Serpentine Visage', 'Uncut Onyx', 'Snakeling', 'Jar of Swamp'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -828,9 +522,7 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Colosseum.',
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 1, max: 3 },
-      hardcore: { min: 2, max: 4 },
     },
   },
   {
@@ -841,9 +533,7 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Fight Caves.',
     quantities: {
-      casual: { min: 1, max: 3 },
       standard: { min: 1, max: 5 },
-      hardcore: { min: 2, max: 7 },
     },
   },
   {
@@ -854,9 +544,7 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Inferno.',
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 1, max: 3 },
-      hardcore: { min: 2, max: 4 },
     },
   },
   {
@@ -867,9 +555,7 @@ const PVM_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Pest Control.',
     quantities: {
-      casual: { min: 7, max: 12 },
       standard: { min: 10, max: 20 },
-      hardcore: { min: 15, max: 30 },
     },
   },
 ];
@@ -883,20 +569,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Alchemical Hydra',
     descriptionTemplate: 'Obtain {quantity} drops from Alchemical Hydra.',
-    acceptableItems: [
-      "Hydra's Tail",
-      'Hydra Leather',
-      "Hydra's Claw",
-      "Hydra's Eye",
-      "Hydra's Fang",
-      "Hydra's Heart",
-      'Ikkle Hydra',
-      'Jar of Chemicals',
-    ],
+    acceptableItems: ['Hydra\'s Tail', 'Hydra Leather', 'Hydra\'s Claw', 'Hydra\'s Eye', 'Hydra\'s Fang', 'Hydra\'s Heart', 'Ikkle Hydra', 'Jar of Chemicals'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -907,18 +582,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Cerberus',
     descriptionTemplate: 'Obtain {quantity} drops from Cerberus.',
-    acceptableItems: [
-      'Primordial Crystal',
-      'Pegasian Crystal',
-      'Eternal Crystal',
-      'Smouldering Stone',
-      'Hellpuppy',
-      'Jar of Souls',
-    ],
+    acceptableItems: ['Primordial Crystal', 'Pegasian Crystal', 'Eternal Crystal', 'Smouldering Stone', 'Hellpuppy', 'Jar of Souls'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 3 },
-      hardcore: { min: 3, max: 4 },
     },
   },
   {
@@ -929,19 +595,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Corporeal Beast',
     descriptionTemplate: 'Obtain {quantity} drops from Corporeal Beast.',
-    acceptableItems: [
-      'Pet Dark Core',
-      'Spectral Sigil',
-      'Arcane Sigil',
-      'Elysian Sigil',
-      'Holy Elixir',
-      'Spirit Shield',
-      'Jar of Spirits',
-    ],
+    acceptableItems: ['Pet Dark Core', 'Spectral Sigil', 'Arcane Sigil', 'Elysian Sigil', 'Holy Elixir', 'Spirit Shield', 'Jar of Spirits'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -952,16 +608,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Corrupted Hunllef',
     descriptionTemplate: 'Obtain {quantity} drops from Corrupted Hunllef.',
-    acceptableItems: [
-      'Crystal Weapon Seed',
-      'Enhanced Crystal Weapon Seed',
-      'Crystal Armour Seed',
-      'Youngllef',
-    ],
+    acceptableItems: ['Crystal Weapon Seed', 'Enhanced Crystal Weapon Seed', 'Crystal Armour Seed', 'Youngllef'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -972,14 +621,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Crystalline Hunllef',
     descriptionTemplate: 'Obtain {quantity} drops from Crystalline Hunllef.',
-    acceptableItems: [
-      'Crystal Weapon Seed',
-      'Enhanced Crystal Weapon Seed',
-      'Crystal Armour Seed',
-      'Youngllef',
-    ],
+    acceptableItems: ['Crystal Weapon Seed', 'Enhanced Crystal Weapon Seed', 'Crystal Armour Seed', 'Youngllef'],
     quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -994,9 +638,7 @@ const PVM_MASTER = [
     descriptionTemplate: 'Obtain {quantity} drops from Doom of Mokhaioitl.',
     acceptableItems: ['Dom', 'Avernic Treads', 'Eye of Ayak', 'Mokhaiotl Cloth'],
     quantities: {
-      casual: { min: 1, max: 1 },
-      standard: { min: 1, max: 2 },
-      hardcore: { min: 2, max: 3 },
+      standard: { min: 1, max: 1 },
     },
   },
   {
@@ -1007,20 +649,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Duke Sucellus',
     descriptionTemplate: 'Obtain {quantity} drops from Duke Sucellus.',
-    acceptableItems: [
-      'Chromium Ingot',
-      'Eye of the Duke',
-      'Virtus Robe Top',
-      'Virtus Mask',
-      'Virtus Robe Bottom',
-      'Magus Ring',
-      'Baron',
-      'Gold Ring',
-    ],
+    acceptableItems: ['Chromium Ingot', 'Eye of the Duke', 'Virtus Robe Top', 'Virtus Mask', 'Virtus Robe Bottom', 'Magus Ring', 'Baron'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -1031,20 +662,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'The Leviathan',
     descriptionTemplate: 'Obtain {quantity} drops from The Leviathan.',
-    acceptableItems: [
-      'Chromium Ingot',
-      'Venator Ring',
-      "Lil'viathan",
-      "Leviathan's Lure",
-      'Virtus Robe Top',
-      'Virtus Mask',
-      'Virtus Robe Bottom',
-      'Gold Ring',
-    ],
+    acceptableItems: ['Chromium Ingot', 'Venator Ring', 'Lil\'viathan', 'Leviathan\'s Lure', 'Virtus Robe Top', 'Virtus Mask', 'Virtus Robe Bottom'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -1055,19 +675,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Nex',
     descriptionTemplate: 'Obtain {quantity} drops from Nex.',
-    acceptableItems: [
-      'Zaryte Vambraces',
-      'Torva Full Helm',
-      'Torva Platebody',
-      'Torva Platelegs',
-      'Ancient Hilt',
-      'Nihil Horn',
-      'Nexling',
-    ],
+    acceptableItems: ['Zaryte Vambraces', 'Torva Full Helm', 'Torva Platebody', 'Torva Platelegs', 'Ancient Hilt', 'Nihil Horn', 'Nexling'],
     quantities: {
-      casual: { min: 1, max: 1 },
       standard: { min: 1, max: 1 },
-      hardcore: { min: 1, max: 2 },
     },
   },
   {
@@ -1078,20 +688,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Nightmare',
     descriptionTemplate: 'Obtain {quantity} drops from Nightmare.',
-    acceptableItems: [
-      "Inquisitor's Mace",
-      "Inquisitor's Great Helm",
-      "Inquisitor's Hauberk",
-      "Inquisitor's Plateskirt",
-      'Nightmare Staff',
-      'Eldritch Orb',
-      'Harmonised Orb',
-      'Volatile Orb',
-      'Little Nightmare',
-      'Jar of Dreams',
-    ],
+    acceptableItems: ['Inquisitor\'s Mace', 'Inquisitor\'s Great Helm', 'Inquisitor\'s Hauberk', 'Inquisitor\'s Plateskirt', 'Nightmare Staff', 'Eldritch Orb', 'Harmonised Orb', 'Volatile Orb', 'Little Nightmare', 'Jar of Dreams'],
     quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -1102,23 +701,11 @@ const PVM_MASTER = [
     type: 'item_collection',
     role: 'PVMER',
     difficulty: 'master',
-    boss: "Phosani's Nightmare",
-    descriptionTemplate: "Obtain {quantity} drops from Phosani's Nightmare.",
-    acceptableItems: [
-      'Parasitic Egg',
-      "Inquisitor's Mace",
-      "Inquisitor's Great Helm",
-      "Inquisitor's Hauberk",
-      "Inquisitor's Plateskirt",
-      'Nightmare Staff',
-      'Eldritch Orb',
-      'Harmonised Orb',
-      'Volatile Orb',
-      'Little Nightmare',
-      'Jar of Dreams',
-    ],
+    boss: 'Phosani\'s Nightmare',
+    descriptionTemplate: 'Obtain {quantity} drops from Phosani\'s Nightmare.',
+    acceptableItems: ['Parasitic Egg', 'Inquisitor\'s Mace', 'Inquisitor\'s Great Helm', 'Inquisitor\'s Hauberk', 'Inquisitor\'s Plateskirt', 'Nightmare Staff', 'Eldritch Orb', 'Harmonised Orb', 'Volatile Orb', 'Little Nightmare', 'Jar of Dreams'],
     quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -1131,20 +718,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Vardorvis',
     descriptionTemplate: 'Obtain {quantity} drops from Vardorvis.',
-    acceptableItems: [
-      'Chromium Ingot',
-      'Ultor Ring',
-      "Executioner's Axe Head",
-      'Butch',
-      'Virtus Robe Top',
-      'Virtus Mask',
-      'Virtus Robe Bottom',
-      'Gold Ring',
-    ],
+    acceptableItems: ['Chromium Ingot', 'Ultor Ring', 'Executioner\'s Axe Head', 'Butch', 'Virtus Robe Top', 'Virtus Mask', 'Virtus Robe Bottom'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -1155,42 +731,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'The Whisperer',
     descriptionTemplate: 'Obtain {quantity} drops from The Whisperer.',
-    acceptableItems: [
-      'Bellator Ring',
-      'Chromium Ingot',
-      'Wisp',
-      'Virtus Robe Top',
-      "Siren's Staff",
-      'Virtus Mask',
-      'Virtus Robe Bottom',
-      'Gold Ring',
-    ],
+    acceptableItems: ['Bellator Ring', 'Chromium Ingot', 'Wisp', 'Virtus Robe Top', 'Siren\'s Staff', 'Virtus Mask', 'Virtus Robe Bottom'],
     quantities: {
-      casual: { min: 1, max: 2 },
       standard: { min: 2, max: 2 },
-      hardcore: { min: 2, max: 3 },
-    },
-  },
-  {
-    id: 'pvm_fortisColosseum',
-    label: 'Sand and Blood',
-    type: 'item_collection',
-    role: 'PVMER',
-    difficulty: 'master',
-    boss: 'Fortis Colosseum',
-    descriptionTemplate: 'Obtain {quantity} drops from the Fortis Colosseum.',
-    acceptableItems: [
-      'Sunfire Fanatic Helm',
-      'Sunfire Fanatic Cuirass',
-      'Sunfire Fanatic Chausses',
-      'Tonalztics of Ralos',
-      'Echo Crystal',
-      'Smol Heredit',
-    ],
-    quantities: {
-      casual: { min: 1, max: 1 },
-      standard: { min: 1, max: 2 },
-      hardcore: { min: 2, max: 3 },
     },
   },
   {
@@ -1201,32 +744,9 @@ const PVM_MASTER = [
     difficulty: 'master',
     boss: 'Yama',
     descriptionTemplate: 'Obtain {quantity} drops from Yama.',
-    acceptableItems: [
-      'Oathplate Helm',
-      'Oathplate Body',
-      'Oathplate Legs',
-      'Soulflame Horn',
-      'Forgotten Lockbox',
-      'Dossier',
-      'Yami',
-    ],
+    acceptableItems: ['Oathplate Helm', 'Oathplate Body', 'Oathplate Legs', 'Soulflame Horn', 'Forgotten Lockbox', 'Dossier', 'Yami'],
     quantities: {
-      casual: { min: 1, max: 1 },
-      standard: { min: 1, max: 2 },
-      hardcore: { min: 2, max: 3 },
-    },
-  },
-  {
-    id: 'pvm_maggotKing',
-    label: 'The Rot Below',
-    type: 'item_collection',
-    role: 'PVMER',
-    difficulty: 'master',
-    boss: 'Maggot King',
-    descriptionTemplate: 'Obtain {quantity} drops from the Maggot King.',
-    acceptableItems: ['Crimson Kisten', 'Elder Venator Fang'],
-    quantities: {
-      casual: { min: 1, max: 1 },
+      casual  : { min: 1, max: 1 },
       standard: { min: 1, max: 2 },
       hardcore: { min: 2, max: 3 },
     },
@@ -1241,45 +761,6 @@ const SKL_INITIATE = [];
 
 const SKL_ADEPT = [
   {
-    id: 'skl_cooking_xp',
-    label: 'Master Chef',
-    type: 'xp_gain',
-    role: 'SKILLER',
-    difficulty: 'adept',
-    descriptionTemplate: 'Earn {quantity} Cooking XP.',
-    quantities: {
-      casual: { min: 500000, max: 1000000 },
-      standard: { min: 1000000, max: 1500000 },
-      hardcore: { min: 1500000, max: 2500000 },
-    },
-  },
-  {
-    id: 'skl_crafting_xp',
-    label: 'Hands of Gold',
-    type: 'xp_gain',
-    role: 'SKILLER',
-    difficulty: 'adept',
-    descriptionTemplate: 'Earn {quantity} Crafting XP.',
-    quantities: {
-      casual: { min: 500000, max: 1000000 },
-      standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
-    },
-  },
-  {
-    id: 'skl_fletching_xp',
-    label: 'Arrowheads and Feathers',
-    type: 'xp_gain',
-    role: 'SKILLER',
-    difficulty: 'adept',
-    descriptionTemplate: 'Earn {quantity} Fletching XP.',
-    quantities: {
-      casual: { min: 250000, max: 500000 },
-      standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
-    },
-  },
-  {
     id: 'skl_agility_xp',
     label: 'Parkour Pro',
     type: 'xp_gain',
@@ -1287,9 +768,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Agility XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1300,22 +779,18 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Firemaking XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
     id: 'skl_fishing_xp',
-    label: "Gone Fishin'",
+    label: 'Gone Fishin\'',
     type: 'xp_gain',
     role: 'SKILLER',
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Fishing XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1326,9 +801,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Herblore XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1339,9 +812,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Hunter XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1352,9 +823,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Mining XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1365,9 +834,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Runecrafting XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1378,9 +845,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Sailing XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1391,9 +856,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Slayer XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1404,9 +867,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Smithing XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1417,9 +878,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Thieving XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1430,9 +889,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Earn {quantity} Woodcutting XP.',
     quantities: {
-      casual: { min: 250000, max: 500000 },
       standard: { min: 500000, max: 1000000 },
-      hardcore: { min: 1000000, max: 2000000 },
     },
   },
   {
@@ -1443,9 +900,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Barbarian Assault.',
     quantities: {
-      casual: { min: 5, max: 7 },
       standard: { min: 7, max: 12 },
-      hardcore: { min: 10, max: 20 },
     },
   },
   {
@@ -1456,9 +911,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Guardians of the Rift.',
     quantities: {
-      casual: { min: 7, max: 12 },
       standard: { min: 10, max: 20 },
-      hardcore: { min: 15, max: 30 },
     },
   },
   {
@@ -1469,9 +922,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Tempoross.',
     quantities: {
-      casual: { min: 7, max: 12 },
       standard: { min: 10, max: 20 },
-      hardcore: { min: 15, max: 30 },
     },
   },
   {
@@ -1482,9 +933,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Wintertodt.',
     quantities: {
-      casual: { min: 7, max: 12 },
       standard: { min: 10, max: 20 },
-      hardcore: { min: 15, max: 30 },
     },
   },
   {
@@ -1495,9 +944,7 @@ const SKL_ADEPT = [
     difficulty: 'adept',
     descriptionTemplate: 'Complete {quantity} Zalcano.',
     quantities: {
-      casual: { min: 7, max: 12 },
       standard: { min: 10, max: 20 },
-      hardcore: { min: 15, max: 30 },
     },
   },
 ];
@@ -1507,13 +954,13 @@ const SKL_MASTER = [];
 const CW_OBJECTIVE_COLLECTIONS = {
   PVMER: {
     initiate: PVM_INITIATE,
-    adept: PVM_ADEPT,
-    master: PVM_MASTER,
+    adept:    PVM_ADEPT,
+    master:   PVM_MASTER,
   },
   SKILLER: {
     initiate: SKL_INITIATE,
-    adept: SKL_ADEPT,
-    master: SKL_MASTER,
+    adept:    SKL_ADEPT,
+    master:   SKL_MASTER,
   },
 };
